Add files via upload
◾In Cell 7, the error message indicates that DecisionTreeClassifier is not recognized. This NameError typically occurs when the necessary library has not been imported, so Python does not recognize the DecisionTreeClassifier class. To resolve this, you need to import the specific library where DecisionTreeClassifier is defined before trying to use it in your code.

In this case, DecisionTreeClassifier is a class within the sklearn.tree module of the scikit-learn library, which is used for various machine learning models. Make sure you've imported DecisionTreeClassifier from this module at the beginning of your code. Check that you have installed scikit-learn in your environment as well; if not, you may need to install it first by running !pip install scikit-learn in a Jupyter Notebook cell or pip install scikit-learn in your terminal.